+++ source/Energinet.DataHub.MeteringPoints.Domain/MasterDataHandling/Components/InvalidCapacityException.cs
@@ -14,6 +14,7 @@
 
 using System;
 using System.Runtime.Serialization;
+using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
 namespace Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components
 {
