+++ source/Energinet.DataHub.MeteringPoints.Domain/MasterDataHandling/Components/MeteringDetails/MeterIdIsNotAllowedRuleError.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.MeteringDetails
+namespace Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components.MeteringDetails
 {
     public class MeterIdIsNotAllowedRuleError : ValidationError
     {
