+++ source/Energinet.DataHub.MeteringPoints.Domain/MasterDataHandling/Components/MeteringDetails/MeterIdIsRequiredRuleError.cs
@@ -14,17 +14,9 @@
 
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.MeteringDetails
+namespace Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components.MeteringDetails
 {
-    public class MeteringMethod : EnumerationType
+    public class MeterIdIsRequiredRuleError : ValidationError
     {
-        public static readonly MeteringMethod Physical = new MeteringMethod(0, nameof(Physical));
-        public static readonly MeteringMethod Virtual = new MeteringMethod(1, nameof(Virtual));
-        public static readonly MeteringMethod Calculated = new MeteringMethod(2, nameof(Calculated));
-
-        private MeteringMethod(int id, string name)
-            : base(id, name)
-        {
-        }
     }
 }
