+++ source/Energinet.DataHub.MeteringPoints.Tests/Domain/MeteringConfigurationTests.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.MeteringPoints.Domain.MeteringDetails;
+using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components.MeteringDetails;
 using Xunit;
 
 namespace Energinet.DataHub.MeteringPoints.Tests.Domain
